Move ShuffleExchange tests into trait
